refacotr(ProfitCalculator): delete type casting

+++ src/main/java/blackjack/domain/betting/ProfitCalculator.java
@@ -27,7 +27,7 @@ public class ProfitCalculator {
     }
 
     public long calculateDealerProfit() {
-        return -(long) playerProfit.values()
+        return -playerProfit.values()
             .stream()
             .mapToLong(profit -> profit)
             .sum();
